requirements.txt
crawl4ai
ipykernel
cssselect
langchain-community
langchain-qdrant
langchain-mistralai
langchain_openai
langchain_ollama
mistralai
fastapi
uvicorn
ddgs